Adding a composer.json and a README to I18n
