+++ src/I18n/composer.json
@@ -1,0 +1,24 @@
+{
+    "name": "cakephp/i18n",
+    "description": "CakePHP Internationalization library with support for messages translation and dates and numbers localization",
+    "license": "MIT",
+    "authors": [
+        {
+        "name": "CakePHP Community",
+        "homepage": "http://cakephp.org"
+    }
+    ],
+    "autoload": {
+        "psr-4": {
+            "Cake\\I18n\\": "."
+        },
+        "files": ["functions.php"]
+    },
+    "require": {
+        "cakephp/core": "dev-master",
+        "ext-intl": "*",
+        "nesbot/Carbon": "~1.13",
+        "aura/intl": "1.1.*"
+    },
+    "minimum-stability": "dev"
+}
