+++ src/Log/README.md
@@ -1,6 +1,6 @@
 # CakePHP Logging Library
 
-The Cache library provides a `Log` service locator for interfacing with
+The Log library provides a `Log` service locator for interfacing with
 multiple logging backends using a simple interface. With the `Log` class it is
 possible to send a single message to multiple logging backends at the same time
 or just a subset of them based on the log level or context.
